Fix clang compilation issues

+++ Zenith/src/PrimitiveRenderer.cpp
@@ -310,7 +310,7 @@ void PrimitiveRenderer::draw_line_custom(const Point2D& from, const Point2D& to,
         assert(y <= end_y);
 
         std::vector<sf::Vertex> vertices;
-        vertices.reserve(end_y - y + 1);
+        vertices.reserve(static_cast<u32>(end_y - y + 1));
 
         for (; y <= end_y; y++)
         {
@@ -332,7 +332,7 @@ void PrimitiveRenderer::draw_line_custom(const Point2D& from, const Point2D& to,
         assert(x <= end_x);
 
         std::vector<sf::Vertex> vertices;
-        vertices.reserve(end_x - x + 1);
+        vertices.reserve(static_cast<u32>(end_x - x + 1));
 
         for (; x <= end_x; x++)
         {
